Agregando relaciones a las entidades
