--- src/main/java/Entities/Partner.java
@@ -1,28 +1,0 @@
-package Entities;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Setter
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name="socio")
-public class Partner {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private String lastName;
-
-    @Column
-    private String address;
-
-}
