+++ src/main/java/com/app/ao/Entities/Book.java
@@ -1,0 +1,48 @@
+package com.app.ao.Entities;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name="libro")
+public class Book {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column
+    private String title;
+
+    @Column
+    private int pages;
+
+    @Column
+    private int units;
+
+    @Column
+    private String isbn;
+
+    @OneToMany(mappedBy = "book", cascade = CascadeType.ALL, fetch = FetchType.LAZY, orphanRemoval = true)
+    private List<Loan> loanList = new ArrayList<>();
+
+    @ManyToOne
+    @JoinColumn(name = "id_autor", nullable = false)
+    private Author author;
+
+    @ManyToOne
+    @JoinColumn(name = "id_editorial", nullable = false)
+    private Editorial editorial;
+
+    @ManyToOne
+    @JoinColumn(name = "id_categoria", nullable = false)
+    private Category category;
+}
