+++ src/main/java/com/app/ao/Entities/Category.java
@@ -1,7 +1,10 @@
-package Entities;
+package com.app.ao.Entities;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Setter
 @Getter
@@ -18,4 +21,7 @@ public class Category {
 
     @Column
     private String name;
+
+    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL, fetch = FetchType.LAZY, orphanRemoval = true)
+    private List<Book> books = new ArrayList<>();
 }
