+++ src/main/java/com/app/ao/Entities/Editorial.java
@@ -1,7 +1,10 @@
-package Entities;
+package com.app.ao.Entities;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Setter
 @Getter
@@ -21,4 +24,7 @@ public class Editorial {
 
     @Column
     private String address;
+
+    @OneToMany(mappedBy = "editorial", cascade = CascadeType.ALL, fetch = FetchType.LAZY, orphanRemoval = true)
+    private List<Book> books = new ArrayList<>();
 }
