+++ src/main/resources/application.properties
@@ -1,14 +1,13 @@
 
 
-#Configuracion fr la base de datos MYSQL
-
+#Configuracion de la base de datos MYSQL
+server.port =8040
 spring.datasource.url=jdbc:mysql://localhost:3308/library_rest
 spring.datasource.username=root
 spring.datasource.password=159753
 
 #Configuracion Hibernate
-
 spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
